requirements.txt
feedparser>=6.0.10
pandas>=1.3.5
python-dateutil
tensorflow>=2.8.3
numpy>=1.21.6
nltk
pymorphy2
re